Next gen devportal configurations 5 - APIPublisherForNewPortal Class
